Factor out index.tsx into home component Set route to automatically redirect to login if not logged in
